requirements.txt
langgraph==1.0.5
langchain==1.2.0
langchain-core==1.2.4
langchain-google-genai==4.1.2
python-dotenv
